+++ lualib/redis.lua
@@ -4,6 +4,7 @@ local socketchannel = require "socketchannel"
 
 local table = table
 local string = string
+local assert = assert
 
 local redis = {}
 local command = {}
@@ -96,41 +97,57 @@ function command:disconnect()
 end
 
 -- msg could be any type of value
-local function pack_value(lines, v)
-	if v == nil then
-		return
-	end
 
-	v = tostring(v)
-
-	table.insert(lines,"$"..#v)
-	table.insert(lines,v)
+local function make_cache(f)
+	return setmetatable({}, {
+		__mode = "kv",
+		__index = f,
+	})
 end
 
+local header_cache = make_cache(function(t,k)
+		local s = "\r\n$" .. k .. "\r\n"
+		t[k] = s
+		return s
+	end)
+
+local command_cache = make_cache(function(t,cmd)
+		local s = "\r\n$"..#cmd.."\r\n"..cmd:upper()
+		t[cmd] = s
+		return s
+	end)
+
+local count_cache = make_cache(function(t,k)
+		local s = "*" .. k
+		t[k] = s
+		return s
+	end)
+
 local function compose_message(cmd, msg)
-	local len = 1
 	local t = type(msg)
+	local lines = {}
 
 	if t == "table" then
-		len = len + #msg
-	elseif t ~= nil then
-		len = len + 1
-	end
-
-	local lines = {"*" .. len}
-	pack_value(lines, cmd)
-
-	if t == "table" then
+		lines[1] = count_cache[#msg+1]
+		lines[2] = command_cache[cmd]
+		local idx = 3
 		for _,v in ipairs(msg) do
-			pack_value(lines, v)
+			v= tostring(v)
+			lines[idx] = header_cache[#v]
+			lines[idx+1] = v
+			idx = idx + 2
 		end
+		lines[idx] = "\r\n"
 	else
-		pack_value(lines, msg)
+		msg = tostring(msg)
+		lines[1] = "*2"
+		lines[2] = command_cache[cmd]
+		lines[3] = header_cache[#msg]
+		lines[4] = msg
+		lines[5] = "\r\n"
 	end
-	table.insert(lines, "")
 
-	local chunk =  table.concat(lines,"\r\n")
-	return chunk
+	return lines
 end
 
 setmetatable(command, { __index = function(t,k)
@@ -159,6 +176,48 @@ end
 function command:sismember(key, value)
 	local fd = self[1]
 	return fd:request(compose_message ("SISMEMBER", {key, value}), read_boolean)
+end
+
+local function compose_table(lines, msg)
+	local tinsert = table.insert
+	tinsert(lines, count_cache[#msg])
+	for _,v in ipairs(msg) do
+		v = tostring(v)
+		tinsert(lines,header_cache[#v])
+		tinsert(lines,v)
+	end
+	tinsert(lines, "\r\n")
+	return lines
+end
+
+function command:pipeline(ops,resp)
+	assert(ops and #ops > 0, "pipeline is null")
+
+	local fd = self[1]
+
+	local cmds = {}
+	for _, cmd in ipairs(ops) do
+		compose_table(cmds, cmd)
+	end
+
+	if resp then
+		return fd:request(cmds, function (fd)
+			for i=1, #ops do
+				local ok, out = read_response(fd)
+				table.insert(resp, {ok = ok, out = out})
+			end
+			return true, resp
+		end)
+	else
+		return fd:request(cmds, function (fd)
+			local ok, out
+			for i=1, #ops do
+				ok, out = read_response(fd)
+			end
+			-- return last response
+			return ok,out
+		end)
+	end
 end
 
 --- watch mode