+++ skynet-src/skynet_timer.h
@@ -5,8 +5,7 @@
 
 int skynet_timeout(uint32_t handle, int time, int session);
 void skynet_updatetime(void);
-uint32_t skynet_gettime(void);
-uint32_t skynet_gettime_fixsec(void);
+uint32_t skynet_starttime(void);
 
 void skynet_timer_init(void);